Fix issue #15
before render a table, we calculate cell size and then split text into
butches (lines) suitable for this size. But it we have empty cell, it will
return empty list of lines and our row render will not be finished. The
pointer will be pointing to the beginning of the page and breaking everything